Revert "f64"
This reverts commit 7cdc4c5b76000256a0bc14d14e270995aac118cc.

+++ src/distributions/multinomial_dist.jl
@@ -2,11 +2,11 @@ using LinearAlgebra
 using Distributions
 
 """
-    multinomial_hyper(α::AbstractArray{Float64,1})
+    multinomial_hyper(α::AbstractArray{Float32,1})
 [Dirichlet Distribution](https://en.wikipedia.org/wiki/Dirichlet_distribution)
 """
 struct multinomial_dist <: distibution_sample
-    α::AbstractArray{Float64,1}
+    α::AbstractArray{Float32,1}
 end
